11_18_2008 PC
git-svn-id: http://dim3.googlecode.com/svn/trunk@320 2c4a6428-2a2a-0410-ad13-7dde452bfe2e

+++ dim3MapUtility/Sources/nodes.c
@@ -95,7 +95,7 @@ int map_find_next_node_in_path(map_type *map,int from_idx,int to_idx)
 {
 	if ((from_idx==-1) || (to_idx==-1)) return(-1);
 	
-	return(map->nodes[from_idx].path_hint[to_idx]);
+	return((int)map->nodes[from_idx].path_hint[to_idx]);
 }
 
 /* =======================================================
